% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.Pengen}
\alias{coef.Pengen}
\title{Obtain Coefficients from Fit}
\usage{
\method{coef}{Pengen}(object, simplify = TRUE, ...)
}
\arguments{
\item{object}{an object of class \code{'Pengen'}.}

\item{simplify}{whether to call \code{\link[=drop]{drop()}} on the result before returning}

\item{...}{arguments that are passed on to \code{\link[stats:update]{stats::update()}}.}
}
\value{
Coefficients from the model after having dropped extraneous
dimensions by calling drop.
}
\description{
This function is equivalent to simply calling \code{drop(object$coefficients)}.
}
\examples{
fit <- golem(mtcars$mpg, mtcars$vs, n_sigma = 1)
coef(fit)
}
