% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fista.R
\name{fista}
\alias{fista}
\title{FISTA}
\usage{
fista(x, y, lambda, family, penalty, intercept, beta, standardize,
  fit_intercept, L = 1, eta = 2, tol_rel_gap = 1e-06,
  tol_infeas = 1e-06, max_passes = 1000)
}
\arguments{
\item{x}{feature matrix}

\item{y}{response}

\item{lambda}{regularization strength}

\item{family}{family object}

\item{penalty}{penalty object}

\item{intercept}{intercept}

\item{beta}{coefficients}

\item{standardize}{whether data is standardized}

\item{fit_intercept}{whether to fit intercept}

\item{L}{Lipschitz constant}

\item{eta}{step size in line search}

\item{tol_rel_gap}{relative tolerance for duality gap check}

\item{tol_infeas}{tolerance for infeasibility check}

\item{max_passes}{max iterations for algorithm}
}
\value{
Coefficients and diagnostics.
}
\description{
FISTA
}
