% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient.R
\name{gradient}
\alias{gradient}
\title{Gradient}
\usage{
gradient(family, linear_predictor, y, x)
}
\arguments{
\item{family}{family}

\item{linear_predictor}{linear predictor}

\item{y}{response}

\item{x}{features}
}
\value{
Gradient vector
}
\description{
Gradient
}
