% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model}
\alias{model}
\title{Fit a prototype for a penalized, generalized linear model}
\usage{
model(x, y, family = "gaussian", penalty = "slope", sigma = NULL,
  lambda = NULL, fdr = 0.2, n_penalties = 100,
  lambda_min_ratio = ifelse(NROW(x) < NCOL(x), 0.01, 1e-04),
  screening = c("none", "strong", "safe"), solver = "fista",
  intercept = TRUE, standardize = TRUE)
}
\arguments{
\item{x}{features}

\item{y}{response}

\item{family}{family object}

\item{penalty}{penalty objet}

\item{sigma}{regularization strength for SLOPE-type penalties}

\item{lambda}{regularization strength for LASSO and SLOPE}

\item{fdr}{target false discovery rate}

\item{n_penalties}{number of penalty terms to apply}

\item{lambda_min_ratio}{smallest size of lambda as a fraction of
lambda_max}

\item{solver}{which solver to use}

\item{intercept}{whether to fit an intercept (bias)}

\item{standardize}{whether to standardize features}
}
\value{
A list of results.
}
\description{
Fit a prototype for a penalized, generalized linear model
}
