% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pengen-package.R
\docType{package}
\name{pengen-package}
\alias{pengen}
\alias{pengen-package}
\title{pengen: Infrastructure for Prototyping Penalized Generalized Linear Models}
\description{
Provides infrastructure for quickly prototyping penalty terms,
    screening rules, and numerical solvers for generalized linear models. 
    The framework is extensible enough to allow users to test out their 
    own models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jolars/pengen}
  \item Report bugs at \url{https://github.com/jolars/pengen/issues}
}

}
\author{
\strong{Maintainer}: Johan Larsson \email{mail@larssonjohan.com} (0000-0002-4029-5945)

}
\keyword{internal}
