% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prox.R
\name{prox}
\alias{prox}
\title{Evaluate proximal operator}
\usage{
prox(object, ...)
}
\arguments{
\item{object}{penalty object, see \link{Penalty}.}

\item{...}{ignored}
}
\value{
Coefficients after evaluating prox
}
\description{
Evaluate proximal operator
}
