% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient.R
\name{pseudoGradient}
\alias{pseudoGradient}
\title{Pseudo-Gradient}
\usage{
pseudoGradient(family, linear_predictor, y)
}
\arguments{
\item{family}{a family, such as \code{\link[=Gaussian]{Gaussian()}}}

\item{linear_predictor}{linear predictor}

\item{y}{response}
}
\value{
Pseudo gradient (needs to be multiplied with X^T)
}
\description{
Compute the gradient (before taking the inner product with the
feature matrix).
}
