% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomProblem.R
\name{randomProblem}
\alias{randomProblem}
\title{Randomly generated problem data}
\usage{
randomProblem(n = 1000, p = 100, q = 0.2, n_groups = NULL,
  density = 1, amplitude = 3, sigma = 1, response = c("gaussian",
  "binomial"))
}
\arguments{
\item{n}{numer of data points}

\item{p}{number of features}

\item{q}{proportion of nonzero features}

\item{n_groups}{number of groups (for group penalties)}

\item{density}{density of features}

\item{amplitude}{scale of coefficients}

\item{sigma}{variability in model}

\item{response}{type of response}
}
\value{
A list of the data along with some other stuff
}
\description{
Randomly generated problem data
}
