% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenFeatures.R
\name{screenFeatures}
\alias{screenFeatures}
\title{Screen features}
\usage{
screenFeatures(family, penalty, lambda, lambda_prev, x, y, beta_prev,
  intercept_prev, method = c("none", "strong", "safe"))
}
\arguments{
\item{family}{family}

\item{penalty}{penalty}

\item{x}{features}

\item{y}{response}

\item{method}{type of screening}

\item{beta}{coefficients from last fit}

\item{intercept}{intercept from last fit}
}
\value{
A logical vector indicating whether to drop a feature or not
}
\description{
Screen features
}
