% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unstandardize.R
\name{unstandardize}
\alias{unstandardize}
\title{Return coefficients to original scale}
\usage{
unstandardize(intercepts, betas, x, y, fit_intercept, x_center, x_scale,
  y_center, y_scale)
}
\arguments{
\item{intercepts}{intercepts}

\item{betas}{coefficients}

\item{x}{features}

\item{y}{response}

\item{fit_intercept}{whether intercept have been fitted}

\item{x_center}{centers of features}

\item{x_scale}{scale of features}

\item{y_center}{centers of responses}

\item{y_scale}{scale of responses}
}
\value{
Coefficients and intercept on original scale.
}
\description{
Return coefficients to original scale
}
