% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient.R
\name{pseudoGradient}
\alias{pseudoGradient}
\title{Compute the "pseudo-gradient"}
\usage{
pseudoGradient(family, ...)
}
\arguments{
\item{family}{a family, such as \code{\link[=Gaussian]{Gaussian()}}}

\item{...}{ignored}
}
\value{
Return the pseudo-gradient.
}
\description{
Compute the gradient (before taking the inner product with the
feature matrix).
}
